refactor balance stream impl design
