+++ mm2src/coins/qrc20.rs
@@ -1486,8 +1486,6 @@ impl MmCoin for Qrc20Coin {
     fn on_disabled(&self) -> Result<(), AbortedError> { AbortableSystem::abort_all(&self.as_ref().abortable_system) }
 
     fn on_token_deactivated(&self, _ticker: &str) {}
-
-    async fn handle_balance_stream(self, _ctx: MmArc) { todo!() }
 }
 
 pub fn qrc20_swap_id(time_lock: u32, secret_hash: &[u8]) -> Vec<u8> {
