+++ mm2src/coins/utxo/bch.rs
@@ -1308,8 +1308,6 @@ impl MmCoin for BchCoin {
             tokens.remove(ticker);
         };
     }
-
-    async fn handle_balance_stream(self, _ctx: MmArc) { todo!() }
 }
 
 impl CoinWithDerivationMethod for BchCoin {
