+++ mm2src/coins/utxo/slp.rs
@@ -1876,8 +1876,6 @@ impl MmCoin for SlpToken {
     fn on_disabled(&self) -> Result<(), AbortedError> { self.conf.abortable_system.abort_all() }
 
     fn on_token_deactivated(&self, _ticker: &str) {}
-
-    async fn handle_balance_stream(self, _ctx: MmArc) { todo!() }
 }
 
 #[async_trait]
