+++ mm2src/coins/utxo/utxo_standard.rs
@@ -793,8 +793,6 @@ impl MmCoin for UtxoStandardCoin {
     fn on_disabled(&self) -> Result<(), AbortedError> { AbortableSystem::abort_all(&self.as_ref().abortable_system) }
 
     fn on_token_deactivated(&self, _ticker: &str) {}
-
-    async fn handle_balance_stream(self, _ctx: MmArc) { todo!() }
 }
 
 #[async_trait]
