+++ mm2src/coins_activation/src/bch_with_tokens_activation.rs
@@ -17,6 +17,7 @@ use common::{drop_mutability, true_f};
 use crypto::CryptoCtxError;
 use mm2_core::mm_ctx::MmArc;
 use mm2_err_handle::prelude::*;
+use mm2_event_stream::EventStreamConfiguration;
 use mm2_number::BigDecimal;
 use serde_derive::{Deserialize, Serialize};
 use serde_json::Value as Json;
@@ -330,4 +331,6 @@ impl PlatformWithTokensActivationOps for BchCoin {
         let settings = AbortSettings::info_on_abort(format!("bch_and_slp_history_loop stopped for {}", self.ticker()));
         self.spawner().spawn_with_settings(fut, settings);
     }
+
+    fn handle_balance_streaming(&self, _config: &EventStreamConfiguration) {}
 }
