+++ mm2src/coins_activation/src/eth_with_token_activation.rs
@@ -15,6 +15,7 @@ use common::Future01CompatExt;
 use common::{drop_mutability, true_f};
 use mm2_core::mm_ctx::MmArc;
 use mm2_err_handle::prelude::*;
+use mm2_event_stream::EventStreamConfiguration;
 #[cfg(target_arch = "wasm32")]
 use mm2_metamask::MetamaskRpcError;
 use mm2_number::BigDecimal;
@@ -277,6 +278,8 @@ impl PlatformWithTokensActivationOps for EthCoin {
         _initial_balance: Option<BigDecimal>,
     ) {
     }
+
+    fn handle_balance_streaming(&self, _config: &EventStreamConfiguration) {}
 }
 
 fn eth_priv_key_build_policy(
