+++ mm2src/coins_activation/src/solana_with_tokens_activation.rs
@@ -18,6 +18,7 @@ use crypto::CryptoCtxError;
 use futures::future::try_join_all;
 use mm2_core::mm_ctx::MmArc;
 use mm2_err_handle::prelude::*;
+use mm2_event_stream::EventStreamConfiguration;
 use mm2_number::BigDecimal;
 use serde_derive::{Deserialize, Serialize};
 use serde_json::Value as Json;
@@ -288,4 +289,6 @@ impl PlatformWithTokensActivationOps for SolanaCoin {
         _initial_balance: Option<BigDecimal>,
     ) {
     }
+
+    fn handle_balance_streaming(&self, _config: &EventStreamConfiguration) {}
 }
