+++ mm2src/mm2_main/src/lp_native_dex.rs
@@ -50,7 +50,6 @@ use crate::mm2::lp_ordermatch::{broadcast_maker_orders_keep_alive_loop, clean_me
 use crate::mm2::lp_swap::{running_swaps_num, swap_kick_starts};
 use crate::mm2::rpc::spawn_rpc;
 
-use coins::coin_balance_event::CoinBalanceEvent;
 use mm2_event_stream::behaviour::EventBehaviour;
 use mm2_net::network_event::NetworkEvent;
 
@@ -392,10 +391,8 @@ fn migration_1(_ctx: &MmArc) {}
 
 fn init_event_streaming(ctx: &MmArc) {
     // This condition only executed if events were enabled in mm2 configuration.
-
     if let Some(config) = &ctx.event_stream_configuration {
         NetworkEvent::new(ctx.clone()).spawn_if_active(config);
-        CoinBalanceEvent::new(ctx.clone()).spawn_if_active(config);
     }
 }
